+++ tap/__init__.py
@@ -127,3 +127,4 @@ def main_service(global_config, **settings):
     return config.make_wsgi_app()
 
 
+__version__ = '0.3.5'