+++ doc/source/CHANGES.rst
@@ -1,6 +1,13 @@
 版本记录
 ========
 
+
+0.3.5
+-----
+- Fix bug: can't get cache data in the page API->CONFIG->TAB[CACHE]
+- Config support: set api-config-cache time, default is 3 minutes, set ENVIRONMENT CONFIG TAP_API_RELOAD to override it.
+
+
 0.3.4
 -----
 - 登录部分增加验证码验证登录